api/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Fiona==1.8.13.post1
Flask==1.1.1
geopandas==0.7.0
idna==2.9
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
munch==2.5.0
numpy==1.18.4
packaging==19.1
pandas==1.0.4
pluggy==0.12.0
postgis==1.0.4
psycopg2-binary==2.8.5
py==1.8.0
pyparsing==2.4.2
pyproj==2.6.1.post1
pytest==5.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
Shapely==1.7.0
six==1.12.0
urllib3==1.25.9
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
